+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 
 # redisio
 
+## Unreleased
+
+- Fix disable recipe service naming for systemd
+
 ## 4.2.0 (2020-09-14)
 
 - New server option 'permissions' to override default (0644) unix permissions on config file